+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Change the `DonationDialog` `href` prop to an `onClick` callback
 * Change the `ArticleDonationBanner` `href` prop to an `onClick` callback
 * Change the `ThinDonationBanner` `href` prop to an `onClick` callback
+* Remove unused `donateText` prop from `ThinDonationBanner` component
 
 ## 1.23.2 (2019-04-24)